netpoll: Add support for hardware checksumming on egress
Netpoll used to compute its own csum; but if the device supports, we
should let it do the checksum itself.

Note: we owe the upstream community some documentation about why and how
we configure FMan for checksumming on egress. It is apparently not clear
to them and so a reason for not having this patch upstream yet.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: Ie2c8210e1d1e7f5be60bcb9610042009548e06bd
Reviewed-on: http://git.am.freescale.net:8181/1539
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>