powerpc/uio: Add uio driver for rapidio message and doorbell units
For using the rapidio message and doorbell units in linux user space,
need to build the uio devices in kernel space. The message and
doorbell uio devices in kernel will map the units registers to user
space and then the applications in user space can configure those
registers and use the rapidio message and doorbell units.

Signed-off-by: Liu Gang <Gang.Liu@freescale.com>
Change-Id: Icfbc6e3cedc627b88143ae1656e6aebc4bd75684
Reviewed-on: http://git.am.freescale.net:8181/981
Reviewed-by: Lian Minghuan-B31939 <Minghuan.Lian@freescale.com>
Reviewed-by: Zang Tiefei-R61911 <tie-fei.zang@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>