Fix geometry configured status check
Checking dma_domain->win_cnt is not sufficient to conclude that
geometry it set.
Either it should check dma_domain->geom_size also or it should
check dma_domain->win_arr as this is allocated when geometry is set.

We are using dma_domain->win_arr pointer check in this patch.

Signed-off-by: Bharat Bhushan <bharat.bhushan@freescale.com>
Change-Id: I67418215ca6e9f495ec80d90a77a15a9cb766d90
Reviewed-on: http://git.am.freescale.net:8181/1487
Reviewed-by: Yoder Stuart-B08248 <stuart.yoder@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>