KVM: PPC: e500mc: Workaround guest missing tlbwe conditional
e6500 has threads but does not have TLB write conditional. Currently HW
threads are exposed to guests as cores so locks are not an option.
Workaround this for guests not allowing one core's threads to use same LPID
simultaneous. The VM's LPID allocation scheme associate one LPID per HW
tread index. This limits the number of VM's to (number of LPIDs - 1)/
(threads per core), which for T4240 processor means 31.

Signed-off-by: Mihai Caraman <mihai.caraman@freescale.com>
Change-Id: Ib1ba743f0f81084decbb26ebcef00d49df814c6c
Reviewed-on: http://git.am.freescale.net:8181/1750
Reviewed-by: Wood Scott-B07421 <scottwood@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>