Check of qemu emulated PCI controller.
While implementing direct device assignment for linux guest using identity
mapped memory (1:1 guest to host physical mapping), we hit a case
where guest end of memory is not a power of two and thus is not
completely mapped by the PCI controller inbound window.

For a similar condition if the PCI controller supports PCI express
capability the fsl_pci driver allows for over commit of inbound
memory window.

In this patch we introduce a check to see if the PCI controller is a
qemu emulated controller. If it's an emulated controller we allow
the inbound memory over commit in the fsl_pci driver.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Change-Id: I6dd7e5fb960a2ec068dd4f01c51906a7a1dea1f7
Reviewed-on: http://git.am.freescale.net:8181/2357
Reviewed-by: Wood Scott-B07421 <scottwood@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>