offline_port: Add support for enabling IP Fragmentation in the O/H port driver
In order to enable fragmentation on an offline port, a chosen node
should be configured in dt for that specific port with two parameters:
manip_extra_space and data_align.
In addition external buffer pools must be set for the O/H port in
the dts file (e.g: fsl,bman-buffer-pools = <&bp8>;).

A bool flag is transmited to dpaa_eth_init_port (used further in FMD wrapper)
macro whether or not fragmentation was enabled.
For the dpaa_eth module this flag must be set to false.(FMD frag support is
only available on O/H ports).

Change-Id: Ifb01b2feebab9199cc0f72b611ce212061d0a7e2
Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2261
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>