gianfar: Add BD rings allocation to L2 SRAM
When this option is enabled the driver will try to allocate the
buffer descriptors (BDs) to the L2 SRAM memory, using the
powerpc/85xx cache-sram support API for the P1/P2 QorIQ platforms.
For this to work, the size and the address of the SRAM region
must be provided as kernel boot params.
Should the attempt to allocate in L2 SRAM fail, the driver falls
back to normal allocation of BDs as before (i.e. to DDR).

The gfar_l2sram_en driver module param is provided to allow to
disable the allocation to L2 SRAM at driver level (enabled by
default).
Sysfs stubs allow to enable/diable allocation of BDs to L2 SRAM
for individual gianfar devices, as the L2 SRAM is an expensive
resource. By default, each device has this feature enabled.
This feature also disables the BD L2 stasing feature by default.
Stasing however may still be configured per device via sysfs.

Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Change-Id: Ie967cd24c2c6d6654b33d175be0d9e3c785c324a
Reviewed-on: http://git.am.freescale.net:8181/2596
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Gupta Rajan-B15745 <rajan.gupta@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>