dpa_offload: Fix for Policer engine in dpa_ipsec_sa_add_policy
When the function dpa_ipsec_sa_add_policy is called, the policy
parameters are copied in an internal structure. The policer
parameters are provided as a pointer so the parameters should
be copied in an internal allocated memory area and not used "as is".

This patch is related to CR ENGR00263529.

Change-Id: I342f0f06457bb128452b33bed3eac55752c4bd6b
Signed-off-by: Anca Jeanina FLOAREA <anca.floarea@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2625
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Varvara Andrei-B21317 <andrei.varvara@freescale.com>
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>