dpa_offload: Remove unused eth8 and eth9 from the B4860QDS dts file
Ethernet interfaces 8 and 9 are usually mapped on TGEC interfaces,
but they are not used by the dpa_offloading applications. These
interfaces were removed from the dts file because if the RCW of the
board is not in sync with the board hardware (like extension cards),
it can cause errors at initialization.

Change-Id: I16aba4d555a66d6e93fbb3f7b0d0e2ac5116060c
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2653
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>