powerpc/dpaa: workaround of fixed-link
fixed-link does not work in 3.8.x, to make XFI work which has no PHY,
it's actually not necessary to use fixed-link, otherwise there will be
more work and code changes to make fixed-link work in 3.8.x. For now, if
driver does not find the phy_node of MAC, just set the mac_dev->phy_dev
= NULL and return 0; we can get XFI work and this won't break other
ethernet ports. It's only for SDK1.4, if we decide to bring fixed-link
back, it can be done after SDK1.4.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Change-Id: I930edbe76f5d739f7d6565d3a7f8a3aabf224db0
Reviewed-on: http://git.am.freescale.net:8181/2797
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zang Tiefei-R61911 <tie-fei.zang@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>