gianfar: Fix tx/rx queue dma_base programming for 36bit mode
For 36 bit mode, TBASEH/RBASEH eTSEC regs need to be programmed
with the upper 32-bit part of the DMA base addr of the BD region.
This is needed when the BDs are allocated in L2 SRAM above the 4G
boundary.
For addresses below 4G, which is the default use case, these
registers just maintain their default reset value of 0.

Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Change-Id: I69e3b86c09ad11780d7124031dd582dc8cd697f0
Reviewed-on: http://git.am.freescale.net:8181/2696
Reviewed-by: Gupta Rajan-B15745 <rajan.gupta@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>