KVM: PPC: Book3e: Get vcpu's last instruction for emulation
lwepx faults needs to be handled by KVM and this implies additional code
in DO_KVM (for DTB_MISS, DSI and LRAT) to check ESR[EPID] and EPLC[EGS]
which is too intrusive for the host. Get rid of lwepx and acquire
last instuction in kvmppc_handle_exit() by searching for the physical address
and kmap it. This fixes an infinite loop caused by lwepx data TLB miss
handled in the host and the TODO for TLB eviction and execute-but-not-read
entries.

Signed-off-by: Mihai Caraman <mihai.caraman@freescale.com>
Change-Id: Id1255d31406a2322292ff06a63c164e861c45fd3
Reviewed-on: http://git.am.freescale.net:8181/2679
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Wood Scott-B07421 <scottwood@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>