fmd: Conditionally set default MAX_FRAME_SIZE
FSL_FM_MAX_FRAME_SIZE used to be conservatively set to 1522, to
accomodate the driver without S/G support. But in fact the default value
should depend on whether S/G support is available or not.
Using 9600 as the default max frame size, if S/G support is available in
the DPAA Ethernet driver.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I82d4c0b41f4a46e3464eb4d64239bb3d4a4158dc
Reviewed-on: http://git.am.freescale.net:8181/2759
Reviewed-by: Radulescu Ruxandra Ioana-B05472 <ruxandra.radulescu@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>