Revert "fmd: Conditionally set default MAX_FRAME_SIZE"
The initial patch was meant to allow the user to set any MTU she desires
without rebooting the system and changing the kernel bootargs.
This however led to problems when the user would manually choose
FSL_DPAA_ETH_OPTIMIZE_FOR_IPFWD, because the default MAX_FRAME_SIZE=9600
would remain selected and this wasn't all that obvious.

Reverting to the previous config selection is the best choice now.

This reverts commit 68dbabbdb1e28169a372530ca3712c3b63296322.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I71499e9b86f8909d049ff4347c4fe035e67d704a
Reviewed-on: http://git.am.freescale.net:8181/2971
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Schmitt Richard-B43082 <B43082@freescale.com>
Tested-by: Schmitt Richard-B43082 <B43082@freescale.com>